Fix a bug where prinfo breaks on unsuccessful SetPrinterInfoList Prior to this change, prinfo would break when the winapi_printer function SetPrinterInfoList failed. The app would just crash and not notfiy the user about what happend.
This change removes a throw statement in the body of the if statement
that checks if SetPrinterInfoList failed. If it failed the program threw
an error and tried to return the empty printerlist afterwards. Throwing
the error is unecessary since the return of an empty printerlist is
enough to understand that probably something went wrong.

Fix #25.