Update and simplify the service functions
Prior to this change, there was only a function to restart the spooler
service. Which is not reusable at all. And the coming function which
clears the printers folder needs to stop and start the spooler service
manually.

This change adds a function to setup the connection to a service with a
given name and adds functions to start and stop this service. The
restart function was refactored accordingly.