Add two new formatting functions
The addition of the printer bins needed two new formatting functions.

This change adds these new functions.
- One to convret an string vector to
a regular string. The elements of the vector are concatenated to a
single string. The elements are seperated by a given seperator string
parameter.
- The other function removes trailing zeros ('\0') from strings. The bin
functions retrieve strings that can have 0 to 23 trailing zeros. This
functions removes the unnecessary zeros.