Add paper bins to printer infos
Prior to this change, the information about the available paper bins and
the default source of a printer were not provided.

This change adds the function setBins() which collects informations
about a printers paper bins. The default output of a printer now
includes informations about the available bins and the default paper
source of a printer.