Add new menu option and navigation
This change adds a new option to the functions menu. It does not
require admin rights and enables the user to export all classes to a csv
file. It's inserted as option 5 in the functions menu.

This change adds a new navigation bar which extends the SaveBackQuit
bar. It adds the option to export printers. It is implemented into
option 1 of the main menu and option 1 of the functions menu.