Add function to read file name and comment
This change adds a function that makes it internally easier and less
redundant to ask for file names and comments. The function takes a
file type and a bool argument. The file struct will hold information
about the type, name and comment. The bool comment determines if the
user should be asked for a comment. Txt for example supports optional
comments, csv does not.