Clear console in ExportPrinters
ExportPrinters can be called from the menu and from the navigation bars.
When its called from the menu the current display works fine because the
menu clears the console before executing the function. But when its
called from a navigation bar it gets printed directly beneath the nav
bar which does not look good.

Now the ExportPrinters function clears the console by itself, reprints
the header and then prints its actual content.